pyproject.toml
[project]
name = "lt-investments"
version = "0.1.0"
description = "Long-term investment alert system that identifies optimal buying opportunities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "alpha-vantage>=2.3.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
lt-investments = "main:main"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "B", "A", "C4", "UP"]
ignore = ["E203", "E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]